Don't error if target line doesn't appear traceable
The idea behind this error was tracing code like this:

    local function increment_sum(counter)
      sum = sum + counter
      -- trace this line
      return sum
    end

Since the "trace this line" line is blank/a comment, Lua's line hook
will never actually reach it, and the tracer will never fire - it seems
reasonable to complain to the user that the line they're trying to trace
will never actually be run.

My current detection implementation just checks if the target
line is between the first and last lines of a function, and if it is,
if the target line is among the function's "active lines".  This works
well, as long as you're not nesting functions:

    local function main()
        local sum = 0

        local function increment_sum(counter)
          sum = sum + counter
          -- trace this line
          return sum
        end

        for i = 1, 10 do
            increment_sum(i)
        end
    end

Here, main's first and last lines surround the target line, but the
target line is also not an active line within main, so it would be
folly to error out here.  I'm not sure if I can implement this (one
idea could be to crawl through the program's functions, looking for
one that has the target line as an active one - but what if that
function ends up getting GC'd?), and I'm not going to stop to think
about the best way to do this right now.